refactor: jwt 기능 구조 통합
