--- src/lib/jwt/auth-jwt/auth-jwt.interface.ts
@@ -1,4 +1,0 @@
-export interface AuthJwtPayload {
-  /** User ID */
-  id: string;
-}
