+++ src/lib/jwt/decorators/temporary-jwt.decorator.ts
@@ -1,6 +1,6 @@
 import { createParamDecorator, ExecutionContext } from '@nestjs/common';
 
-import { TemporaryJwtPayload } from './temporary-jwt.interface';
+import { TemporaryJwtPayload } from '../jwt-utility.interface';
 
 export const TemporaryUserAuth = createParamDecorator((_: unknown, ctx: ExecutionContext) => {
   const request = ctx.switchToHttp().getRequest<Request & { user: TemporaryJwtPayload }>();
